add login and logout service
